First import: Add EasyMesh for simple 2d mesh generation.
EasyMesh is written by Bojan Niceno (niceno@wt.tn.tudelft.nl) and the licence says "This program is distributed with no warranty of any kind. The author does not accept any responsability for the consequences of using it, or for wheather it serves any particular purpose or works at all. EasyMesh is freeware, and you may use it for any purpose at no cost, with no need for any licence. You may use it at home or at your place of work, for private or for business purposes. You may not distribute EasyMesh as a component of any commercial product without a prior license agreement with the author." (see download.html).

See <http://www-dinma.univ.trieste.it/nirftc/research/easymesh/> for more information.